iex/requirements.txt
aiohttp==3.8.1
aiometer==0.3.0
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
